requirements.txt
filterpy==1.4.5
lap==0.5.12
matplotlib==3.10.8
numpy==2.4.0
opencv_contrib_python==4.11.0.86
pandas==2.3.3
Pillow==12.0.0
Requests==2.32.5
scipy==1.16.3
seaborn==0.13.2
skimage==0.0
torch==2.2.2
torchvision==0.17.2
ultralytics==8.3.223
